requirements.txt
asgiref==3.11.0
attrs==25.4.0
Django==4.2.27
django-jazzmin==3.0.1
djangorestframework==3.16.1
drf-spectacular==0.29.0
gunicorn==23.0.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
packaging==26.0
PyYAML==6.0.3
referencing==0.36.2
rpds-py==0.27.1
sqlparse==0.5.5
typing_extensions==4.15.0
uritemplate==4.2.0
